Added ability to store strings to document

+++ store.sh
@@ -6,8 +6,8 @@ PARENT=""
 
 # Set Date for today
 Define_Today() {
-	read -r YEAR MONTH DAY <<<$(date +%Y\ %b\ %d)
-	DATE="$YEAR-$MONTH-$DAY"
+	read -r YEAR MONTH DAY MONTH_NUM <<<$(date +%Y\ %b\ %d\ %m)
+	DATE="$YEAR-$MONTH_NUM-$DAY"
 }
 
 # Set Time for now
@@ -39,8 +39,13 @@ Write_LabBook() {
     fi
 
     echo "\\subsection*{${TIME}} " >> $FILENAME
-    cat $INFILE >> $FILENAME
-    rm $INFILE
+	if [ $PLAINTEXT ]
+	then
+		echo ${INFILE} >> $FILENAME
+	else
+    	cat $INFILE >> $FILENAME
+		rm $INFILE
+	fi
 }
 
 Append_Image() {
@@ -67,9 +72,10 @@ fi
 LABFILE=""
 IMAGEFILE=""
 IMAGESIZE=""
+PLAINTEXT=false
 
 # Captures the user flags
-while getopts "hL:C:I:s:" opt; do
+while getopts "hL:T:C:I:s:" opt; do
 	case $opt in
 		h)
 			echo "this message will be helpful in the future"
@@ -77,6 +83,10 @@ while getopts "hL:C:I:s:" opt; do
 			;;
 		L)
 			LABFILE=$OPTARG
+			;;
+		T)
+			LABFILE=$OPTARG
+			PLAINTEXT=true
 			;;
 		C)
 			echo "This is not implemented yet"
